portfolio-tracker/portfolio-tracker-pro/backend/requirements-dev.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
plotly==5.17.0
python-binance==1.0.19
ccxt==4.1.77
tabulate==0.9.0
reportlab==4.0.7
pybit==5.12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
slowapi==0.1.9
prometheus-client==0.20.0
redis==5.0.8
scikit-learn==1.3.2
prophet==1.1.5
numpy>=1.26.0
ta==0.11.0
torch==2.2.0
transformers==4.35.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

